Move loading the ORT configuration to the model
This allows to reuse the code to load the configuration in helper-cli.
