+++ model/src/main/kotlin/config/OrtConfiguration.kt
@@ -19,6 +19,14 @@
 
 package com.here.ort.model.config
 
+import com.here.ort.utils.expandTilde
+
+import com.typesafe.config.ConfigFactory
+
+import io.github.config4k.extract
+
+import java.io.File
+
 /**
  * The configuration model for all ORT components.
  */
@@ -27,4 +35,29 @@ data class OrtConfiguration(
      * The configuration of the scanner.
      */
     val scanner: ScannerConfiguration? = null
-)
+) {
+    companion object {
+        /**
+         * Load the [OrtConfiguration]. The different sources are used with this priority:
+         *
+         * 1. [Command line arguments][args]
+         * 2. [Configuration file][configFile]
+         * 3. default.conf from resources
+         */
+        fun load(args: Map<String, String> = emptyMap(), configFile: File? = null): OrtConfiguration {
+            val argsConfig = ConfigFactory.parseMap(args, "Command line").withOnlyPath("ort")
+            val fileConfig = configFile?.expandTilde()?.let {
+                ConfigFactory.parseFile(it).withOnlyPath("ort")
+            }
+            val defaultConfig = ConfigFactory.parseResources("default.conf")
+
+            var combinedConfig = argsConfig
+            if (fileConfig != null) {
+                combinedConfig = combinedConfig.withFallback(fileConfig)
+            }
+            combinedConfig = combinedConfig.withFallback(defaultConfig).resolve()
+
+            return combinedConfig.extract("ort")
+        }
+    }
+}
